requirement.txt
google-generativeai
pandas 
numpy
matplotlib
notebook
python-dotenv
openai
langchain
librosa
pydub